pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mixedbread-ai-haystack"
version = "2.1.1"
description = "Mixedbread Haystack integration"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
keywords = ["haystack", "mixedbread-ai", "embedding", "reranking", "parsing", "nlp", "llm"]
authors = [{ name = "Mixedbread AI", email = "support@mixedbread.com" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "haystack-ai>=2.0.0",
    "mixedbread-ai>=0.3.1"
]

[project.urls]
Homepage = "https://mixedbread.com"
Documentation = "https://github.com/mixedbread-ai/mixedbread-ai-haystack#readme"
Issues = "https://github.com/mixedbread-ai/mixedbread-ai-haystack/issues"
Source = "https://github.com/mixedbread-ai/mixedbread-ai-haystack"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["mixedbread_ai_haystack*"]
exclude = ["data*", "tests*", "examples*"]

[tool.pytest.ini_options]
markers = [
    "integration: integration tests that call the Mixedbread AI API (deselect with '-m \"not integration\"')",
]
log_cli = true

